+++ backend/src/neuroagent/app/routers/qa.py
@@ -1,5 +1,6 @@
 """Endpoints for agent's question answering pipeline."""
 
+import asyncio
 import json
 import logging
 from contextlib import asynccontextmanager
@@ -18,6 +19,7 @@ from obp_accounting_sdk import AsyncAccountingSessionFactory
 from obp_accounting_sdk.constants import ServiceSubtype
 from openai import AsyncOpenAI
 from redis import asyncio as aioredis
+from semantic_router import SemanticRouter
 
 from neuroagent.agent_routine import AgentsRoutine
 from neuroagent.app.app_utils import (
@@ -37,6 +39,7 @@ from neuroagent.app.dependencies import (
     get_context_variables,
     get_openai_client,
     get_redis_client,
+    get_semantic_routes,
     get_settings,
     get_starting_agent,
     get_thread,
@@ -141,6 +144,7 @@ async def stream_chat_agent(
     accounting_session_factory: Annotated[
         AsyncAccountingSessionFactory, Depends(get_accounting_session_factory)
     ],
+    semantic_router: Annotated[SemanticRouter | None, Depends(get_semantic_routes)],
     background_tasks: BackgroundTasks,
 ) -> StreamingResponse:
     """Run a single agent query in a streamed fashion."""
@@ -206,6 +210,41 @@ async def stream_chat_agent(
         proj_id=thread.project_id,
         count=1,
     ):
+        if semantic_router:
+            selected_route = await semantic_router.acall(user_request.content)
+            if selected_route.name:  # type: ignore
+                # If a predefined route is detected, return predefined response
+                async def yield_predefined_response(
+                    response: str,
+                ) -> AsyncIterator[str]:
+                    """Imitate the LLM streaming."""
+                    for chunk in response.split(" "):
+                        await asyncio.sleep(0.01)
+                        yield f"0:{json.dumps(chunk + ' ', separators=(',', ':'))}\n"
+
+                response = next(
+                    route.metadata["response"]  # type: ignore
+                    for route in semantic_router.routes
+                    if route.name == selected_route.name  # type: ignore
+                )
+                messages.append(
+                    Messages(
+                        order=len(messages),
+                        thread_id=thread.thread_id,
+                        entity=Entity.AI_MESSAGE,
+                        content=json.dumps({"role": "assistant", "content": response}),
+                        is_complete=True,
+                    )
+                )
+                return StreamingResponse(
+                    yield_predefined_response(response),
+                    media_type="text/event-stream",
+                    headers={
+                        "x-vercel-ai-data-stream": "v1",
+                        **limit_headers.model_dump(by_alias=True),
+                    },
+                )
+
         stream_generator = stream_agent_response(
             agents_routine=agents_routine,
             agent=agent,